+++ Project1/UnitCommandable.h
@@ -20,12 +20,12 @@ public:
 
 	// Inventory
 	void add(std::unique_ptr<Item> item);
-	std::unique_ptr<Item> drop(int item_index);
+	std::unique_ptr<Item> drop(size_t item_index);
 	std::unique_ptr<Item> drop(Item& item);
-	void swap(int i, int j);
+	void swap(size_t i, size_t j);
 	void swap(Item& a, Item& b);
 
-	void equip(EquipSlot slot, int item_index);
+	void equip(EquipSlot slot, size_t item_index);
 	void equip(EquipSlot slot, Item& item);
 	void equip(EquipSlot slot, std::unique_ptr<Item> item);
 	void dequip(EquipSlot slot);
